Move reporting views to their own file.
