+++ zproject/urls.py
@@ -128,10 +128,10 @@ urlpatterns += patterns('zerver.views',
     url(r'^json/tutorial_status$',          'json_tutorial_status'),
     url(r'^json/change_enter_sends$',       'user_settings.json_change_enter_sends'),
     url(r'^json/get_profile$',              'json_get_profile'),
-    url(r'^json/report_error$',             'json_report_error'),
-    url(r'^json/report_send_time$',         'json_report_send_time'),
-    url(r'^json/report_narrow_time$',       'json_report_narrow_time'),
-    url(r'^json/report_unnarrow_time$',     'json_report_unnarrow_time'),
+    url(r'^json/report_error$',             'report.json_report_error'),
+    url(r'^json/report_send_time$',         'report.json_report_send_time'),
+    url(r'^json/report_narrow_time$',       'report.json_report_narrow_time'),
+    url(r'^json/report_unnarrow_time$',     'report.json_report_unnarrow_time'),
     url(r'^json/update_message_flags$',     'messages.json_update_flags'),
     url(r'^json/register$',                 'json_events_register'),
     url(r'^json/upload_file$',              'json_upload_file'),
